requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.8
lxml==5.3.0
oauth2client==4.1.3
openai==1.42.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
simplegmail==4.1.1
six==1.16.0
soupsieve==2.6
uritemplate==4.1.1
urllib3==2.2.2
